TR-1284 Renamed RouteContext to RouterContext
